+++ src/SFA.DAS.ProviderRelationships.Api.Client/ReadStore/Data/IAccountProviderLegalEntitiesRepository.cs
@@ -2,7 +2,6 @@ using SFA.DAS.ProviderRelationships.Api.Client.ReadStore.Models;
 
 namespace SFA.DAS.ProviderRelationships.Api.Client.ReadStore.Data
 {
-//    internal interface IAccountProviderLegalEntitiesRepository : IReadOnlyDocumentRepository<IAccountProviderLegalEntityDto>
     internal interface IAccountProviderLegalEntitiesRepository : IReadOnlyDocumentRepository<AccountProviderLegalEntityDto>
     {
     }
