+++ src/SFA.DAS.ProviderRelationships.Api.Client/ReadStore/Models/AccountProviderLegalEntityDto.cs
@@ -1,27 +1,15 @@
 using System;
 using System.Collections.Generic;
 using Newtonsoft.Json;
-using SFA.DAS.CosmosDb;
 using SFA.DAS.ProviderRelationships.Types.Models;
 
 namespace SFA.DAS.ProviderRelationships.Api.Client.ReadStore.Models
 {
-    public interface IAccountProviderLegalEntityDto
-    {
-        long AccountLegalEntityId { get; }
-
-        long Ukprn { get; }
-
-        IEnumerable<Operation> Operations { get; }
-
-        DateTime? Deleted { get; }
-    }
-
     //todo: make these internal if poss
     //todo: interface so the original and copy don't get out of sync. interface in types
     //todo: if use interface, then should be able to fakeiteasy
     //todo: could we plug in a CamelCaseNamingStrategy? see https://stackoverflow.com/questions/21783137/default-camel-case-of-property-names-in-json-serialization
-    public class AccountProviderLegalEntityDto : IAccountProviderLegalEntityDto //: IDocument
+    public class AccountProviderLegalEntityDto : IAccountProviderLegalEntityDto
     {
         [JsonProperty("accountLegalEntityId")]
         public virtual long AccountLegalEntityId { get; private set; }
@@ -34,10 +22,5 @@ namespace SFA.DAS.ProviderRelationships.Api.Client.ReadStore.Models
 
         [JsonProperty("deleted")]
         public virtual DateTime? Deleted { get; private set; }
-
-//        [JsonConstructor]
-//        public AccountProviderLegalEntityDto()
-//        {
-//        }
     }
 }
